Team registration and team list. - New messages handled: ClubRegisterTeam, ClubUnregisterTeam and ClubTeamList. - Removed unneeded appendResult method from MessageUtils.

+++ src/main/java/com/neikeq/kicksemu/game/rooms/messages/RoomMessages.java
@@ -169,7 +169,7 @@ public class RoomMessages {
     }
 
     public static void leaveRoom(Session session, ClientMessage msg) {
-        short roomId = msg.readShort();
+        int roomId = msg.readShort();
         int playerId = session.getPlayerId();
 
         Room room = RoomManager.getRoomById(roomId);
